Refactored trpc code and removed utils folder
