+++ src/components/navigations/Header.tsx
@@ -9,7 +9,7 @@ import LogoDark from '~/components/logos/LogoDark';
 import { buttonVariants } from '~/components/primitives/Button';
 import { ThemeToggle } from '~/components/utility/ThemeToggle';
 import { api } from '~/lib/api';
-import { Roles } from '~/utils/constants';
+import { Roles } from '~/lib/constants';
 import { Icons } from '../icons/Icons';
 import HeaderLink from '../links/HeaderLink';
