+++ src/lib/printConsoleLogo.ts
