+++ src/server/api/procedures/adminProcedure.ts
@@ -1,0 +1,6 @@
+import { isAdmin } from '../middlewares/authMiddleware';
+import { t } from '../trpc';
+
+const adminProcedure = t.procedure.use(isAdmin);
+
+export default adminProcedure;
