+++ src/server/api/procedures/protectedProcedure.ts
@@ -1,0 +1,6 @@
+import { isAuthed } from '../middlewares/authMiddleware';
+import { t } from '../trpc';
+
+const protectedProcedure = t.procedure.use(isAuthed);
+
+export default protectedProcedure;
