+++ src/server/api/procedures/publicProcedure.ts
@@ -1,0 +1,5 @@
+import { t } from '../trpc';
+
+const publicProcedure = t.procedure;
+
+export default publicProcedure;
