+++ src/server/api/routers/pet.ts
@@ -1,7 +1,8 @@
-import { createTRPCRouter, publicProcedure } from "~/server/api/trpc";
+import { createTRPCRouter } from '~/server/api/trpc';
+import publicProcedure from '../procedures/publicProcedure';
 
 export const petRouter = createTRPCRouter({
   getAllPets: publicProcedure.query(() => {
-    return "will return all pets";
+    return 'will return all pets';
   }),
 });
